[#24] 재료 리스트 화면 로직 구현
- 재료 리스트 화면 UiState 추가
- 기존 Model 제거

+++ app/src/main/java/com/hyeok/recipebook/presentation/ingredient/component/IngredientCard.kt
@@ -28,7 +28,7 @@ import com.hyeok.recipebook.designsystem.theme.YorinTheme
 @Composable
 fun IngredientCard(
     ingredientName: String,
-    remainExpirationDate: Long,
+    remainExpirationDate: Int,
     expirationDate: String,
     weight: Int,
     weightUnit: String,
